Updated train_critic.py

+++ mt/erac/train_critic.py
@@ -229,7 +229,7 @@ def evaluate(iterator):
         Q_all = critic(tgt, seq, out_mode=models.LOGIT)
 
         # compute Q(y_{<t}, y_t)
-        Q_mod = Q_all.gather(2, seq[1:].unsqueeze(2)).squeeze(2)                # [tgtlen-1 x bsz]
+        Q_mod = Q_all.gather(2, seq[1:].unsqueeze(2).to(torch.int64)).squeeze(2)                # [tgtlen-1 x bsz]
 
         # compute V_hat(y_{<t})
         act_log_dist = act_log_dist.data.clone()
